Create README.md
The Project is based on the front end on Django Server to test the product and theme. I have used Bootstrap for Design and Implementation to match the requirements of the Page. This is my first time using Django for Front-end use and I get a chance to become familiar with HTML and CSS. The theme is based on the Product Page.